usecase-snowflake/requirements.txt
apache-beam==2.19.0
asn1crypto==1.3.0
avro-python3==1.9.2.1
azure-common==1.1.25
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
boto3==1.11.17
botocore==1.14.17
cachetools==3.1.1
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
crcmod==1.7
cryptography==2.9
dill==0.3.1.1
docopt==0.6.2
docutils==0.15.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-api-core==1.16.0
google-apitools==0.5.28
google-auth==1.11.2
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
hdfs==2.5.8
httplib2==0.12.0
idna==2.9
ijson==2.6.1
jmespath==0.9.5
mock==2.0.0
monotonic==1.5
numpy==1.18.1
oauth2client==3.0.0
oscrypto==1.2.0
pbr==5.4.4
protobuf==3.11.3
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodomex==3.9.7
pydot==1.4.1
PyJWT==1.7.1
pymongo==3.10.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
snowflake-connector-python==2.2.4
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
